Move integration test to test_comms.py
