+++ distributed/comm/tests/test_comms.py
@@ -55,6 +55,17 @@ def tcp(monkeypatch, request):
     return tcp
 
 
+@pytest.fixture(params=["tornado"])
+def uds(monkeypatch, request):
+    """Set the TCP backend to either tornado or asyncio"""
+    if request.param == "tornado":
+        import distributed.comm.uds as uds
+    else:
+        raise NotImplementedError()
+    monkeypatch.setitem(backends, "uds", uds.UDSBackend())
+    return uds
+
+
 ca_file = get_cert("tls-ca-cert.pem")
 
 # The Subject field of our test certs
@@ -100,10 +111,6 @@ async def get_comm_pair(listen_addr, listen_args=None, connect_args=None, **kwar
 
 def get_tcp_comm_pair(**kwargs):
     return get_comm_pair("tcp://", **kwargs)
-
-
-def get_uds_comm_pair(**kwargs):
-    return get_comm_pair("uds://", **kwargs)
 
 
 def get_tls_comm_pair(**kwargs):
@@ -197,7 +204,7 @@ def test_resolve_address(tcp):
     assert f("tcp://localhost:456") == "tcp://127.0.0.1:456"
     assert f("tls://localhost:456") == "tls://127.0.0.1:456"
 
-    assert f("uds://%2Ftmp%2Fdask.sock") == "uds://localhost"
+    assert f("uds://%2Ftmp%2Fdask.sock") == "uds://%2Ftmp%2Fdask.sock"
 
 
 def test_get_local_address_for(tcp):
@@ -212,8 +219,6 @@ def test_get_local_address_for(tcp):
     inproc_res = f(inproc_arg)
     assert inproc_res.startswith("inproc://")
     assert inproc_res != inproc_arg
-
-    assert f("uds://%2Ftmp%2Fdask.sock") == "uds://%2Ftmp%2Fdask.sock"
 
 
 #
@@ -284,6 +289,52 @@ async def test_tcp_specific(tcp):
     futures = [client_communicate(key=i, delay=0.05) for i in range(N)]
     await asyncio.gather(*futures)
     assert set(l) == {1234} | set(range(N))
+
+
+@gen_test()
+async def test_uds_specific(uds):
+    """
+    Test concrete UDS API.
+    """
+
+    async def handle_comm(comm):
+        assert comm.peer_address == (f"uds://{host}:0")
+        assert comm.extra_info == {}
+        msg = await comm.read()
+        msg["op"] = "pong"
+        await comm.write(msg)
+        await comm.close()
+
+    listener = await uds.UDSListener("localhost", handle_comm)
+    host, port = listener.get_host_port()
+
+    assert host.endswith(".sock")
+    assert port == 0  # we fake port 0 when using UDS
+
+    l = []
+
+    async def client_communicate(key, delay=0):
+        comm = await connect(listener.contact_address)
+        assert comm.peer_address == f"uds://{host}:0"
+        assert comm.extra_info == {}
+        await comm.write({"op": "ping", "data": key})
+        if delay:
+            await asyncio.sleep(delay)
+        msg = await comm.read()
+        assert msg == {"op": "pong", "data": key}
+        l.append(key)
+        await comm.close()
+
+    await client_communicate(key=1234)
+
+    # Many clients at once
+    N = 100
+    futures = [client_communicate(key=i, delay=0.05) for i in range(N)]
+    await asyncio.gather(*futures)
+    assert set(l) == {1234} | set(range(N))
+
+    listener.stop()
+    assert not os.path.exists(host)  # assert socket deleted
 
 
 @pytest.mark.parametrize("sni", [None, "localhost"])
